+++ src/rdx/reducer/index.ts
@@ -1,0 +1,8 @@
+import { gameField } from './gameField';
+import { nextMove } from './nextMove';
+import { combineReducers } from 'redux';
+
+export const reducer = combineReducers({
+  nextMove,
+  gameField
+})
